formatting

+++ src/infinity/interfaces/IToucan.sol
@@ -3,13 +3,13 @@
 pragma solidity ^0.8.16;
 
 interface IToucanPool {
-    function redeemAuto2(uint amount) external returns (address[] memory tco2s, uint[] memory amounts);
+    function redeemAuto2(uint256 amount) external returns (address[] memory tco2s, uint256[] memory amounts);
 
-    function redeemMany(address[] calldata erc20s, uint[] calldata amounts) external;
+    function redeemMany(address[] calldata erc20s, uint256[] calldata amounts) external;
 
-    function feeRedeemPercentageInBase() external pure returns (uint);
+    function feeRedeemPercentageInBase() external pure returns (uint256);
 
-    function feeRedeemDivider() external pure returns (uint);
+    function feeRedeemDivider() external pure returns (uint256);
 
     function redeemFeeExemptedAddresses(address) external view returns (bool);
 
@@ -17,14 +17,14 @@ interface IToucanPool {
 }
 
 interface IToucanCarbonOffsets {
-    function retire(uint amount) external;
+    function retire(uint256 amount) external;
 
     function retireAndMintCertificate(
         string calldata retiringEntityString,
         address beneficiary,
         string calldata beneficiaryString,
         string calldata retirementMessage,
-        uint amount
+        uint256 amount
     ) external;
 
     function mintCertificateLegacy(
@@ -32,7 +32,7 @@ interface IToucanCarbonOffsets {
         address beneficiary,
         string calldata beneficiaryString,
         string calldata retirementMessage,
-        uint amount
+        uint256 amount
     ) external;
 }
